+++ src/main/java/com/github/games647/lagmonitor/commands/ThreadCommand.java
@@ -11,6 +11,9 @@ import org.bukkit.command.CommandSender;
 
 public class ThreadCommand implements CommandExecutor {
 
+    private static final ChatColor PRIMARY_COLOR = ChatColor.DARK_AQUA;
+    private static final ChatColor SECONDARY_COLOR = ChatColor.DARK_GREEN;
+
     private final LagMonitor plugin;
 
     public ThreadCommand(LagMonitor plugin) {
@@ -26,7 +29,8 @@ public class ThreadCommand implements CommandExecutor {
                 continue;
             }
 
-            sender.sendMessage(ChatColor.GOLD + thread.getName() + " State: " + thread.getState());
+            sender.sendMessage(PRIMARY_COLOR + thread.getName()
+                    + ChatColor.GOLD + " State: " + SECONDARY_COLOR + thread.getState());
         }
 
         return true;